Add Echo client and server example
This adds the echo example which demonstrate a few ways to use
the Mojo C++ bindings to bounce strings around mojo messages. This
is composed of two client apps, echo_client and echo_client_sync,
that demonstrate how to use the synchronous and asynchronous bindings
and a server echo_server that demonstrates a few flavors of serving
logic possible.

The echo_benchmark client is disabled as it depends on various bits
of base (like tracing and string to number conversion) that aren't in
our world yet.

Change-Id: I3a55cb1c6129a6edb13ad1bb6d052373712f6b67