Add dart:mojo.internals package
This package is linked into the dart_content_handler and into the
flutter_content_handler.

This location matches the location of this code in domokit/mojo. Over time, we
might want to consider moving it elsewhere, but matching the domokit location
for now will make it easier to bring up code in Fuchsia.

Change-Id: Ie9618e0b0a8e7196947737cd5292861a4f8d69cb